puzzel15
In the game app known as Fifteen of Game) 15 Puzzle. (Game Description:
The game board consists of 16 slots, arranged in 4x4 (four rows and four columns).
1 to 15 are on the board, each number appears only once, and one box is empty)
The beginning of the game (.
At any point in the game, you can move a slot near the empty slot) that is above, below, left or right
The empty slot (only. The adjacent slot will take up the empty slot).
Purpose of the Game - Arrange the numbers in ascending order, with the number 1 appearing in the top left slot, the number